[inductor] Refactor pre-grad passes into inductor.fx_passes
[ghstack-poisoned]
