+++ torch/_inductor/compile_fx.py
@@ -30,6 +30,7 @@ from .debug import DebugContext
 from .decomposition import select_decomp_table
 from .fx_passes.joint_graph import joint_graph_passes
 from .fx_passes.post_grad import post_grad_passes
+from .fx_passes.pre_grad import pre_grad_passes
 from .graph import GraphLowering
 from .utils import (
     developer_warning,
@@ -654,9 +655,7 @@ def compile_fx(
 
         # Since handle_dynamo_export_graph will trigger compile_fx again,
         # Move these passes after handle_dynamo_export_graph to avoid repeated calls.
-        with overrides.patch_functions():
-            model_ = overrides.replace_fx(model_, example_inputs_)
-            model_ = overrides.fuse_fx(model_, example_inputs_)
+        pre_grad_passes(model_, example_inputs_)
 
     if any(isinstance(x, (list, tuple, dict)) for x in example_inputs_):
         return flatten_graph_inputs(
